Add deploy button on the dashboard page and refractor dashboard page
