+++ client/src/components/dashboard/webpage/WebpageTable.js
@@ -1,0 +1,20 @@
+import {
+    Container,
+    Table,
+    Button,
+    Card,
+    Modal,
+    DropdownButton,
+    Dropdown,
+  } from 'react-bootstrap';
+import React, {useState} from 'react'; 
+
+const WebPageTable = ({currentWebPages}) => {
+    return(
+        <div>
+            
+        </div>
+    )
+}
+
+export default WebPageTable
