+++ client/src/components/layouts/DashboardSidebarData.js
@@ -45,11 +45,6 @@ const DashboardSidebarData = () => {
       name: 'TEAM',
     },
     {
-      title: 'Deploy Website',
-      icon: <BsUpload />,
-      link: '/deploy',
-    },
-    {
       title: 'Settings',
       icon: <BsGearFill />,
       link: '/dashboard/settings',
