+++ client/src/components/route/PrivateRoute.js
@@ -44,7 +44,6 @@ const PrivateRoute = () => {
             component={PublicationPage}
           />
           <DashboardLayoutRoute path="/team" exact component={TeamPage} />
-          <DashboardLayoutRoute path="/deploy" exact component={DeployPage} />
           {/*  If login, any other route not stated above will be redirect dashbroad page */}
           <Redirect to="/dashboard"/> 
         </Switch>
